Refactor how plugin name and id are imported
