--- admin/src/pluginId.js
@@ -1,5 +1,0 @@
-const pluginPkg = require( '../../package.json' );
-
-const pluginId = pluginPkg.name.replace( /^@strapi\/plugin-/i, '' );
-
-module.exports = pluginId;
