+++ admin/src/utils/index.js
@@ -1,1 +1,3 @@
 export { default as getTrad } from './get-trad';
+export { default as pluginId } from './plugin-id';
+export { default as pluginName } from './plugin-name';
