+++ admin/src/utils/plugin-name.js
@@ -1,0 +1,3 @@
+import pluginPkg from '../../../package.json';
+
+export default pluginPkg.strapi.displayName;
