+++ server/services/preview-button.js
@@ -3,10 +3,11 @@
 const qs = require( 'qs' );
 
 const config = require( '../config' );
+const { pluginId } = require( '../utils' );
 
 module.exports = ( { strapi } ) => ( {
   async getConfig() {
-    const data = await strapi.config.get( 'plugin.preview-button', config.default );
+    const data = await strapi.config.get( `plugin.${pluginId}`, config.default );
 
     return data;
   },
