+++ server/utils/index.js
@@ -1,7 +1,9 @@
 'use strict';
 
 const getService = require( './get-service' );
+const pluginId = require( './plugin-id' );
 
 module.exports = {
   getService,
+  pluginId,
 };
