+++ server/utils/plugin-id.js
@@ -1,0 +1,3 @@
+const pluginPkg = require( '../../package.json' );
+
+module.exports = pluginPkg.strapi.name;
